# api/artists/model_create_artist_request.go
/*
 * Bakeoff API
 *
 * Simple REST API
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package artists

type CreateArtistRequest struct {
	Name string `json:"name"`

	Genre string `json:"genre"`
}