# api/artists/model_update_artist_request.go
/*
 * Bakeoff API
 *
 * Simple REST API
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package artists

type UpdateArtistRequest struct {
	Name string `json:"name"`

	Genre string `json:"genre"`
}